requirements.txt
aiohttp==3.11.18
flask==3.1.1
python-dotenv==1.1.0
psutil==7.0.0
gunicorn==23.0.0